chore: add new forced mode components to system
