+++ packages/components/system/package.json
@@ -45,6 +45,7 @@
     "react": ">=18"
   },
   "devDependencies": {
+    "@chakra-ui/color-mode": "workspace:*",
     "@chakra-ui/theme": "workspace:*",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
